// src/app/CMS/shared/utilities/validation-rules.enum.ts
export enum ValidationRules {
  AlphanumericOnly = 'AlphanumericOnly',
  NumberOnly = 'NumberOnly',
  LettersOnly = 'LettersOnly',
  LettersWithWhiteSpace = 'LettersWithWhiteSpace',
  DecimalOnly = 'DecimalOnly',
  EmailID = 'EmailID',
  Pincode = 'Pincode',
  IFSCCode = 'IFSCCode',
  Address = 'Address',
  MobileNoWithSeries = 'MobileNoWithSeries',
  AgeAbove18Years = 'AgeAbove18Years',
  AccountNumber = 'AccountNumber',
  HouseOrSurveyNumber = 'HouseOrSurveyNumber',
  MobileWithCountryCode = 'MobileWithCountryCode',
  NoSpecialChars = 'NoSpecialChars',
  PasswordStrength = 'PasswordStrength'
}